refactor: Group common attrs in Event/Vis. [DHIS2-13618]
